+++ integration/airflow/openlineage/lineage_backend/__init__.py
@@ -7,8 +7,6 @@ import uuid
 from typing import Optional
 
 from airflow.lineage.backend import LineageBackend
-from airflow.version import version as AIRFLOW_VERSION
-from pkg_resources import parse_version
 
 
 class Backend:
@@ -93,7 +91,8 @@ class OpenLineageBackend(LineageBackend):
     @classmethod
     def send_lineage(cls, *args, **kwargs):
         # Do not use LineageBackend approach when we can use plugins
-        if parse_version(AIRFLOW_VERSION) >= parse_version("2.3.0.dev0"):
+        from openlineage.airflow.utils import is_airflow_version_enough
+        if is_airflow_version_enough("2.3.0"):
             return
         # Make this method a noop if OPENLINEAGE_DISABLED is set to true
         if os.getenv("OPENLINEAGE_DISABLED", None) in [True, 'true', "True"]: