+++ integration/airflow/tests/integration/tests/airflow/dags/postgres_orders_popular_day_of_week.py
@@ -1,12 +1,11 @@
 # Copyright 2018-2022 contributors to the OpenLineage project
 # SPDX-License-Identifier: Apache-2.0
 from airflow import DAG
-from airflow.version import version as AIRFLOW_VERSION
 from airflow.providers.postgres.operators.postgres import PostgresOperator
 from airflow.utils.dates import days_ago
 
+from openlineage.airflow.utils import is_airflow_version_enough
 from openlineage.client import set_producer
-from pkg_resources import parse_version
 set_producer("https://github.com/OpenLineage/OpenLineage/tree/0.0.1/integration/airflow")
 
 
@@ -29,8 +28,6 @@ default_args = {
 }
 
 
-
-
 dag = DAG(
     'postgres_orders_popular_day_of_week',
     schedule_interval='@once',
@@ -41,8 +38,8 @@ dag = DAG(
     description='Determines the popular day of week orders are placed.'
 )
 
-if parse_version(AIRFLOW_VERSION) < parse_version('2.5.0'):
-    t1 = PostgresOperator(    
+if not is_airflow_version_enough('2.5.0'):
+    t1 = PostgresOperator(
         task_id='postgres_if_not_exists',
         postgres_conn_id='food_delivery_db',
         sql="{{ get_sql() }}",